Add BaseTree.isLog(Block) and further fix global purge.
Added BaseTree.isLog(Block) for use in places where what is desired is
knowing whether or not a block is something we should recognize as a
log.

BaseTree.getTreeTypeByBlock(Block) seems to be the most comprehensive
way to determine whether a block is a log, so the implementation of
isLog(Block) is based on a call to getTreeTypeByBlock(Block).

Also fixed the global purge condition so it uses this new isLog(Block)
method. Now it won't remove entries for mushroom trees and custom trees.