Populate more of source with BaseTree.isLog(Block)
Found some other instances where a call to BaseTree.isLog(Block) would
probably be better than a compound condition that would have to be
changed in multiple places later.

I found other candidates for replacement as well, but wasn't sure
replacing them with isLog(Block) was appropriate, as isLog(Block)
returns true in the case of a huge mushroom block. Example: isLog(Block)
might not be appropriate for the check in
TreeAssistBlockListener#calcAir(Block), unless it's okay for leaves not
to decay quickly if they are near a huge mushroom block, so I left that
condition alone.