Revamp FlatFileBlockList to use ConfigurationSection
Addressing performance bottlenecks upon loading and saving by using
Bukkit's configuration framework for in-memory storage instead of just a
middleman.

I would have preferred using a HashSet for the blocklist, but I know of
no mechanism in the Bukkit configuration framework that handles saving
and loading those, which means we're stuck with the same loading and
saving bottleneck as before.

One can achieve performance similar to a HashMap using Bukkit's
configuration framework through a shallow ConfigurationSection.

ConfigurationSections use Strings as keys.

The keys used in this code are identification strings derived from the
coordinates and world of the TreeBlock. I considered using a string of
the hashCode, but there is a risk of collision with that (collision:
multiple TreeBlocks that would not be considered equal but that do have
the same hashCode). The coordinates and world name provide a reasonable
degree of uniqueness to each TreeBlock and form a suitably unique
identification string.

Loading times have been improved dramatically by this change. I would
expect saving times have too.

I believe ditching the Bukkit configuration framework in favor of some
kind of binary storage file is the only way to
really optimize the performance, but that requires more recoding and
then makes the file format much less friendly to work with if you're a
server admin.

I have a different blocklist that I hacked together a few months ago,
but it's over 700 lines of Sid insanity mixed with slipcor insanity, and
I'm not ready to hack it up to the current version of TreeAssist yet. So
I've hacked this together instead.
Seems livable.

Oh! And this way, it's still possible for you to add more information to
TreeBlock later if you want to. I considered just having it write the
timestamp as the value for each key, basically doing away with
TreeBlock, but it occurrs to me maybe there are plans to add more data
to TreeBlock, so I kept the TreeBlock as value. It makes for redundant
data in the yml file and in memory, but it preserves flexibility and it
makes my changes less radical than they could have been.